Remove plugin neoterm to avoid undefined variable
The neoterm plugin sets the 'g:neoterm_repl_python' variable to ""
before ftdetect file 'set_repl_cmd.vim' uses it. The dein plugin
optimises the ftdetect script into a cached ftdetect file that
executes before the plugin, and so the 'g:neoterm_repl_python'
variable is not initialised. This results in an E121 ("Undefined
variable") error and a subsequent cascade of errors.