Catch E676 error caused by vim-gnupg plugin
Nvim gives E676 error ('No matching commands for acwrite buffer') even
though there ARE matching autocommands (see augroup command in
syntax.vim), so catch and ignore this error.

This error occurred when editing '*.gpg' files with the vim-gnupg plugin
enabled, because the plugin sets the buffer type to 'acwrite'.

Vim not checked to see whether it also gives this error in the same
situation.