Fix error with s:plugin_dir
When a sourced file calls a function from the parent script, any script
variable it contains is resolved in the context of the sourced file. In
the case of VrcPluginDir, it was intended that the script variable used
would be from the parent script. As a result, the function had to be
rewritten to not use a script variable.