vim: if markdown check for pandoc-xnos filters
The markdown support function now provides feedback if pandoc or any of
the pandoc-xnos filters are not available. It became apparent the
function was being run multiple times when opening vim to edit an
existing markdown file -- the error messages were being displayed three
times. So, added a variable that is true on the first execution and
false on subsequent executions. This variable is used to ensure feedback